Move lodepng library into core
