+++ platform/Linux/Makefile
@@ -16,7 +16,7 @@ TARGET := $(OUT_DIR)/msx1pq_cli
 
 SOURCES := \
 $(CLI_DIR)/msx1pq_cli.cpp \
-$(CLI_DIR)/lodepng.cpp \
+$(CORE_DIR)/lodepng.cpp \
 $(CORE_DIR)/MSX1PQCore.cpp \
 $(CORE_DIR)/MSX1PQOutput.cpp \
 $(CORE_DIR)/MSX1PQPalettes.cpp
