+++ src/core/lodepng.cpp
