+++ src/core/lodepng.h
